feat(group_v2): add serde of handshake peer info exchange packet of group chat v2
