+++ src/toxcore/messenger/packet/messenger/nickname.rs
