+++ src/toxcore/messenger/packet/messenger/typing.rs
